Adding a parent page that will consume this bluedialog component
